:package: chore(api) migrate learning content repo to `src`
