+++ api/src/learning-content/domain/services/learning-content-conversion-service.js
